// openapi/models/event-status.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Skrib
 * Skrib
 *
 * The version of the OpenAPI document: 2.11.7
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @enum {string}
 */

export const EventStatus = {
    Draft: 'DRAFT',
    Scheduled: 'SCHEDULED',
    InProgress: 'IN_PROGRESS',
    Preflight: 'PREFLIGHT',
    Complete: 'COMPLETE',
    Cancelled: 'CANCELLED',
    Archive: 'ARCHIVE',
    ArchiveComplete: 'ARCHIVE_COMPLETE'
} as const;

export type EventStatus = typeof EventStatus[keyof typeof EventStatus];